Import hengband-1.6.2 from wip
Hengband is a rogue-like game which is a descendant of Moria and
Angband, and is a direct derivation of Zangband. In order to win the
game by defeating The Serpent of Chaos who lurks in the 100th floor of
the Dungeon of Angband, you need to not only obtain a good set of
gears, but also acquaint yourself with the game.